chore: Extract actions related code from BookingListItem.
This commit is easier to review if you look at the diff between
BookingListItem and BookingListItemActions directly. Basically, the code
removed from BookingListItem are refactored into BookingListItemActions.

Ideally, the dialog-related code should be refactored away as well.
